geolocator/requirements.txt
azure-functions
geopy